Convert the c field of nfa_state_t to a 256 byte array.
This is future work to allow a single state to match multiple alphabets.
This will be useful when implementing . operator as well as character
classes. It may also be useful to optimize the state machines to combine
expressions such as a|b|c into a single state.

PS: surprisingly changing the type of `c' from char to a 256 byte array
as improved the performance on the benchmark