Optimize alternation of two characters in a pattern.
Usually the OR (|) operator results in a state machine with 3 nodes, a
null (or epsilon) transition node connected to the two nodes matching
the two alphabets in parallel, this results in two parallel paths in the
state machine. The simplest optimization is two reduce these 3 nodes
into a single node.

Although, it's a good idea but requires more work to be effective.
For example in a pattern of the form a|b|c|d, we have 4 alphabets
Or'ed and this could be reduced to a single state. But the above
optimization will work only for c|d, thereafter it is not possible
for the compiler to know that the right side node of `b` is also an OR
node.