Optimize AST for ? and + operators
An expression of type a? can be convered into NULL | a
Similarly a+ can be converted into aa*. Doing this while building
the AST results in eliminating the complexity in the compiler which
needed to do plumbing of states to create loops. This also cuts down the
number of states needed in the compiled state machine.

This has cut down the timing on the benchmark expression
a?^na^n from around 1000 microseconds to around 63 microseconds for n = 100