Simplifying freeing up the nfa.
Since the NFA is like a cyclic graph - it is not easy to free all the
states in a single iteration - because some nodes are pointed to by
multiple nodes and we may end up visiting them multiple times. The
earlier mechanism was to maintain a list of all the states created and
iterate through the list to free all of them. This had the cost of
upfront memory allocation for the array list as well as resizing and
copying the list over whenever it fell short. The new change avoids this
extra work by delegating it all to the time when we actually need to
free the machine. We do this by iterating through the graph, copying
states into an array, indexed by the state_idx. Later on the list can be
iterated and all states can be freed.